WEB/MallInOne/requirements.txt
dj-database-url==0.5.0
Django==1.11.2
django-crispy-forms==1.7.2
djangorestframework==3.8.2
gunicorn==19.7.1
Markdown==2.6.11
Pillow==5.1.0
psycopg2==2.7.4
Pygments==2.2.0
pytz==2018.4
